snap/snaplock,cmd: move snaplock to cmd/
Samuele asked to move this package to cmd because it is an
implementation detail and not an intrinsics property of a snap.
