+++ cmd/snaplock/lock.go
